clarify the debug scenario
The original text was ambiguous as the user just added a Debug node and there is a sidebar for that node called "debug". There also, it turns out, is a Debug sidebar (vs the Info sidebar).  I was injecting and injecting and nothing was happening...

The doc might be a bit better if it included the actual icon in the text but I did not see that in the `images` folder.